createListener for modal fix
