+++ package.json
@@ -9,7 +9,7 @@
     "ws": "^8.5.0"
   },
   "name": "devland.js",
-  "version": "1.3.4",
+  "version": "1.3.5",
   "description": "A powerful API wrapper for the Discord API designed for modern bots",
   "main": "index.js",
   "types": "./structures/types/index.d.ts",
