+++ structures/models/Collector.js
@@ -53,7 +53,7 @@ module.exports = class Collector extends EventEmitter {
                 if (this.ended) return;
                 if (this.channel && this.channel.id !== interaction.channelId) return;
                 if (this.message && this.message.id !== interaction.message?.id) return;
-                if (typeof this.data.componentType === "number" && interaction.data?.component_type !== this.data.componentType) return;
+                if (typeof this.data.componentType === "number" && (interaction.data?.component_type !== this.data.componentType) && interaction.type !== this.data.componentType) return;
                 if (this.data.filter && typeof this.data.filter === "function") {
                     let array = []
                     array.push(interaction)
